Add support for a new parameter `graphite_flatten_hostname` that converts the hostname into one single graphite metric component.
In our case we do not use the components of the host fqdn as part of the
metric tree. Instead, the fqdn is a single component with dots replaced
by underscores. e.g. foo.example.com becomes foo_example_com.

This commit adds support for this but leaves it turned off by default.
It also adds the missing parameter docs to the class header.