Do *not* use the titleized name of the metric because it can contain spaces.
The titleized name of a metric is its name with underscores replaced
with spaces and the first word capitalized. This causes metrics for
resources with underscores in the name not to work because graphite
doesn't allow spaces in metric names.

This commit changes the report processor to use the standard name instead.